fix(GalleryItemTradesTable.vue): counter swaps not working from gallery item page
