+++ components/trade/TradeActivityTable.vue
@@ -76,7 +76,7 @@
             :trade="item"
             :target="tabTarget"
             :variant="variant"
-            @counter-swap="() => onCounterSwapClick(item)"
+            @counter-swap="() => counterSwap(item)"
             @select="() => {
               selectedTrade = item
               isTradeModalOpen = true
@@ -99,11 +99,7 @@
 
 <script lang="ts" setup>
 import { NeoButton } from '@kodadot1/brick'
-import type {
-  TradeType,
-  Swap,
-  TradeNftItem,
-} from '@/components/trade/types'
+import type { TradeType, TradeNftItem } from '@/components/trade/types'
 
 type TradeTabType = 'outgoing' | 'incoming'
 export type TradeTableQuery = Record<TradeTabType, string>
@@ -114,7 +110,6 @@ const props = defineProps<{
 }>()
 
 const route = useRoute()
-const swapStore = useAtomicSwapStore()
 const { replaceUrl } = useReplaceUrl()
 
 const dataKey = TRADES_QUERY_MAP[props.type].dataKey
@@ -156,32 +151,6 @@ const where = computed(() => {
 })
 
 const { items: trades, loading: loadingTrades } = useTrades({ where, disabled: computed(() => !Object.keys(where.value).length), type: props.type })
-
-const onCounterSwapClick = (trade: TradeNftItem) => {
-  if (!trade.desired) {
-    return
-  }
-
-  const withFields: CrateSwapWithFields = {
-    desired: [tradeToSwapItem(trade.offered)],
-    offered: [tradeToSwapItem(trade.desired)],
-  }
-
-  const tSwap = trade as TradeNftItem<Swap>
-
-  if (tSwap.surcharge) {
-    Object.assign(withFields, {
-      surcharge: {
-        amount: tSwap.price,
-        direction: tSwap.surcharge,
-      },
-    })
-  }
-
-  const swap = swapStore.createSwap(trade.caller, withFields)
-
-  navigateToSwap(swap)
-}
 
 watch(activeTab, value => replaceUrl({ filter: value }))
